requirements.txt
accelerate
anthropic
bitsandbytes
cohere
datasets
evaluate
fire
google-generativeai
ipython
litellm
loguru
nltk
openai
pynvml
rank_bm25
rouge_score
scipy
seaborn
sentence-transformers
tenacity
torch
transformers
umap-learn
vllm
